requirements.txt
click>=6.7,<8
pathos==0.2.2.1
boto3>=1.9.2,<3
botocore>=1.12.204
numpy>=1.18.1,<2 
tenacity>=5,<7
google-cloud-storage>=1,<2
pillow>=6,<8
lz4>=3,<4
tensorflow>=1.14,<2
torch>=1,<2
waymo-open-dataset>=1.0.1,<2